+++ config/routes.rb
@@ -60,6 +60,7 @@ Rails.application.routes.draw do
     get "service_autocomplete", to: "services#autocomplete", as: :service_autocomplete
     get "services/c/:category_id" => "services#index", as: :category_services
     resources :research_areas
+      get "research-areas/popup", to: "research_areas#popup"
     resources :categories
     resources :providers
     resources :platforms